+++ test/vissen/output/context/circle_test.rb
@@ -5,7 +5,7 @@ require 'test_helper'
 describe Vissen::Output::Context::Circle do
   subject { Vissen::Output::Context::Circle }
 
-  let(:radius)         { rand(0.1..1.0) }
+  let(:radius)         { rand(0.1..0.5) }
   let(:point_count)    { 4 }
   let(:circle_context) { subject.new radius, point_count }
 
@@ -13,17 +13,20 @@ describe Vissen::Output::Context::Circle do
     it 'accepts a radius' do
       points = circle_context.points
 
-      assert_in_delta radius,  points[0].x
-      assert_in_delta 0,       points[0].y
+      x0 = circle_context.width / 2
+      y0 = circle_context.height / 2
 
-      assert_in_delta 0,       points[1].x
-      assert_in_delta radius,  points[1].y
+      assert_in_delta x0 + radius, points[0].x
+      assert_in_delta y0 + 0,      points[0].y
 
-      assert_in_delta(-radius, points[2].x)
-      assert_in_delta 0,       points[2].y
+      assert_in_delta x0 + 0,      points[1].x
+      assert_in_delta y0 + radius, points[1].y
 
-      assert_in_delta 0,       points[3].x
-      assert_in_delta(-radius, points[3].y)
+      assert_in_delta(x0 - radius, points[2].x)
+      assert_in_delta y0 + 0,      points[2].y
+
+      assert_in_delta x0 + 0,      points[3].x
+      assert_in_delta(y0 - radius, points[3].y)
     end
 
     it 'accepts a point_count' do
@@ -33,10 +36,41 @@ describe Vissen::Output::Context::Circle do
     it 'accepts an optional angle offset' do
       offset = rand(0...(2 * Math::PI))
       circle_context = subject.new radius, point_count, offset: offset
+
+      x0 = circle_context.width / 2
+      y0 = circle_context.height / 2
+
       points = circle_context.points
 
-      assert_in_delta radius * Math.cos(offset), points[0].x
-      assert_in_delta radius * Math.sin(offset), points[0].y
+      assert_in_delta x0 + radius * Math.cos(offset), points[0].x
+      assert_in_delta y0 + radius * Math.sin(offset), points[0].y
+    end
+
+    it 'accepts a width and a height' do
+      width  = 10
+      height = 5
+      radius = 2.5
+
+      circle_context = subject.new radius, point_count, width: width,
+                                                        height: height
+
+      points = circle_context.points
+      radius_normalized = 1.0 / width * radius
+
+      x0 = circle_context.width / 2
+      y0 = circle_context.height / 2
+
+      assert_in_delta x0 + radius_normalized, points[0].x
+      assert_in_delta y0 + 0,                 points[0].y
+
+      assert_in_delta x0 + 0,                 points[1].x
+      assert_in_delta y0 + radius_normalized, points[1].y
+
+      assert_in_delta(x0 - radius_normalized, points[2].x)
+      assert_in_delta y0 + 0,                 points[2].y
+
+      assert_in_delta x0 + 0,                 points[3].x
+      assert_in_delta(y0 - radius_normalized, points[3].y)
     end
   end
 end